+++ packages/shopify/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @weaverse/shopify
 
+## 3.2.7
+
+### Patch Changes
+
+- @weaverse/react@3.2.7
+
 ## 3.2.6
 
 ### Patch Changes